+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Fix in update settings that needed to restart after updated
 * Correction in the use of the api with mongodb
 * Adjustments to search endpoint for contacts, chats, messages and Status messages
+* Now when deleting the instance, the data referring to it in mongodb is also deleted
 
 # 1.3.1 (2023-07-20 07:48)